Major stability improvements, bug fixes

+++ libc/read.c
@@ -1,11 +1,12 @@
-#include <stdio.h>
 #include <sys/syscall.h>
 #include <sys/defs.h>
+#include <stdio.h>
 
-DEFN_SYSCALL3(read, 2, int, char *, uint16_t);
+DEFN_SYSCALL3(read, 2, int, char *, size_t);
 
-int read(int fd, char *s, uint16_t size)
+ssize_t read(int fd, char *s, size_t size)
 {
 	int bytes = syscall_read(0, s, size);
 	return bytes;
 }
+
